Add a progress bar
