+++ app/src/main/res/layout/quiz_view_pager_fragment.xml
@@ -19,6 +19,14 @@
             layout="@layout/quiz_toolbar"
             bind:category="@{category}"/>
 
+        <ProgressBar
+            android:id="@+id/quiz_timer_progress_bar"
+            style="@style/Widget.AppCompat.ProgressBar.Horizontal"
+            android:max="100"
+            android:progress="0"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"/>
+
         <androidx.viewpager2.widget.ViewPager2
             android:id="@+id/view_pager"
             android:layout_width="match_parent"
